+++ packages/mobile-app/data/facades/README.md
@@ -1,0 +1,3 @@
+# facades
+
+Facades use the `data-facade` package to provide a `tanstack/react-query` interface between the wallet code in the `data` folder and the Expo frontend in the `app` folder.
